Fix bridging and ObjC imports
