+++ vincerp/Util/ReactivePropertyGenerator.swift
@@ -96,15 +96,3 @@ class PropertyObserver: NSObject {
         self.targetObject.removeObserver(self, forKeyPath: propertyName, context: &propertyObserverContext)
     }
 }
-
-public extension UIResponder {
-    
-    public func reactiveProperty<T:Equatable>(forProperty propertyName: String, initValue: T, initializer: ((Source<T>) -> ())? = nil) -> Rx<T> {
-        return ReactivePropertyGeneator.property(self, propertyName: propertyName, initValue: initValue, initializer: initializer)
-    }
-    
-    public func reactiveEmitter<T:Equatable>(name propertyName: String, initValue: T) -> Source<T> {
-        return ReactivePropertyGeneator.emitter(self, propertyName: propertyName, initValue: initValue)
-    }
-    
-}
