+++ vincerp/Util/VRPNSObjectHelper.m
@@ -6,7 +6,6 @@
 #import "VRPNSObjectHelper.h"
 
 #import <objc/runtime.h>
-#import <UIKit/UIKit.h>
 
 @implementation VRPNSObjectHelper
