+++ vincerp/vincerp-Mac.h
@@ -16,4 +16,4 @@ FOUNDATION_EXPORT const unsigned char vincerp_MacVersionString[];
 
 // In this header, you should import all the public headers of your framework using statements like #import <vincerp_Mac/PublicHeader.h>
 
-
+#import "VRPNSObjectHelper.h"
